fix: modifying device power display issues

+++ pos_android_studio_demo/pos_android_studio_app/src/main/java/com/dspread/demoui/activity/MyQposClass.java
@@ -316,7 +316,7 @@ public class MyQposClass extends CQPOSService {
         Intent batteryStatus = getApplicationInstance.registerReceiver(null, ifilter);
         int level = batteryStatus.getIntExtra(BatteryManager.EXTRA_LEVEL, -1);
         int scale = batteryStatus.getIntExtra(BatteryManager.EXTRA_SCALE, -1);
-        int batteryPct = level / scale * 100;
+        float batteryPct = level / (float)scale * 100;
 
         content += getString(R.string.battery_percentage) + batteryPct +"%"+ "\n";
 //			}
